+++ NEWS.md
@@ -1,8 +1,9 @@
-# NVIdb 0.14.1.9000 - (2025-##-##)
+# NVIdb 0.15.0.9001 - (2025-##-##)
 
 ## New features:
 
--
+- `add_MT_omrader` now accepts the argument `shortname`. A short version of the 
+MT unit name will be generated when `shortname = TRUE`.
 
 
 ## Bug fixes:
@@ -20,6 +21,25 @@
 -
 
 
+# NVIdb 0.15.0 - (2025-12-03)
+
+## New features:
+
+- Created `remove_objects`, a convenience wrapper around `rm` | `remove`. The 
+function accepts a vector with named objects and checks if the objects exist 
+before trying to remove them.
+
+- Changed to `MT_sekjonnr` and `MT_seksjon` for `read_MT_omrader` and 
+`add_MT_omrader` in accordance with the terms used by NFSA.
+
+
+## BREAKING CHANGES:
+
+- Changed the name of the last level in the output from `read_MT_omrader` to 
+`MT_sekjonnr` and `MT_seksjon` from `MT_enhetnr` and `MT_enhet`. This also 
+affects the use of `add_MT_omrader`.
+
+
 # NVIdb 0.14.1 - (2025-10-29)
 
 ## New features:
@@ -29,7 +49,7 @@
 
 ## Bug fixes:
 
-- corrected set_dir_NVI for use on workbench.
+- corrected `set_dir_NVI` for use on workbench.
 
 
 ## BREAKING CHANGES:
@@ -48,11 +68,11 @@
 
 ## New features:
 
-- Created read_avlsgris to read the register of nucleus and multiplier pig herds.
+- Created `read_avlsgris` to read the register of nucleus and multiplier pig herds.
 
-- Created read_purkering, to read the register of sow pools herds, i.e. the central herd and the satellite herds.
+- Created `read_purkering`, to read the register of sow pools herds, i.e. the central herd and the satellite herds.
 
-- Created find_file_near_date that finds the file nearest to a wanted date based on the date included in the file name. 
+- Created `find_file_near_date` that finds the file nearest to a wanted date based on the date included in the file name. 
 
 
 ## BREAKING CHANGES: